requirements.txt
fastapi[all]
uvicorn
sqlalchemy
asyncpg
aiosqlite
httpx
apscheduler
pydantic
nats-py